src/dags/empty.py
import datetime

from airflow import DAG
from airflow.operators.bash import BashOperator
from airflow.operators.empty import EmptyOperator

with DAG("my_dag",
    start_date=datetime.datetime(2023, 8, 5),
    schedule="@daily",
    default_args={"retries": 2},
):
    op = BashOperator(task_id="hello_world", bash_command="ls -l")
    print(op.retries)