requirements.txt
Django==1.11.17
celery==4.1.1
geocoder==1.27.0
boto3==1.6.16
gunicorn==20.1.0
redis==2.10.6
xmltodict==0.11.0
kombu==4.2.0
#psycopg2==2.8
apache-airflow==2.6.3
urllib3==1.26.6
boto3
arrow
lxml
rq